Functions to 1) Swap bits in a number. 2) Revers bits in a number.

+++ utilities.h
@@ -1,7 +1,7 @@
 #ifndef UTILITIES_H
 #define UTILITIES_H
 
-void printBinary(unsigned long x);
+void printBinary(unsigned long long x);
 typedef unsigned short ushort;
 
 #endif // UTILITIES_H
